# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeMind AI - AI Trading Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all trades
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, closed, all]
          required: false
        - in: query
          name: asset
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Open a new paper trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a specific trade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade (close position)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
...
  /replay:
    get:
      operationId: listReplays
      tags: [replay]
      summary: List trade replays
      responses:
        "200":
          description: Trade replays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeReplay"
    post:
      operationId: createReplay
      tags: [replay]
      summary: Generate a trade replay for a journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplayRequest"
      responses:
        "201":
          description: Replay created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeReplay"
  /replay/{id}:
    get:
      operationId: getReplay
      tags: [replay]
      summary: Get a specific trade replay
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade replay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeReplay"

  # ── Chat ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Trade:
      type: object
      required: [id, asset, direction, entryPrice, quantity, status, entryReason, marketConditions, aiRecommendation, userAction, createdAt]
      properties:
        id:
          type: integer
        asset:
          type: string
        direction:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        quantity:
          type: number
        status:
          type: string
          enum: [open, closed]
        entryReason:
          type: string
        exitReason:
          type: ["string", "null"]
        marketConditions:
          type: string
        aiRecommendation:
          type: string
          enum: [buy, sell, wait, hold]
        userAction:
          type: string
          enum: [followed, ignored, partial]
        pnl:
          type: ["number", "null"]
        pnlPercent:
          type: ["number", "null"]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        createdAt:
          type: string
        closedAt:
          type: ["string", "null"]
    TradeInput:
      type: object
      required: [asset, direction, entryPrice, quantity, entryReason, marketConditions, aiRecommendation, userAction]
      properties:
        asset:
          type: string
        direction:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        quantity:
          type: number
        entryReason:
          type: string
        marketConditions:
          type: string
        aiRecommendation:
          type: string
          enum: [buy, sell, wait, hold]
        userAction:
          type: string
          enum: [followed, ignored, partial]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
    TradeUpdate:
      type: object
      properties:
        exitPrice:
          type: number
        exitReason:
          type: string
        status:
          type: string
          enum: [closed]
...
    TradeReplay:
      type: object
      required: [id, tradeId, asset, entryDate, entryPrice, exitPrice, direction, marketContext, newsEvents, sentimentConditions, whaleActivity, technicalIndicators, userReasoning, aiRecommendation, actualOutcome, lessonsLearned, pnl, pnlPercent, createdAt]
      properties:
        id:
          type: integer
        tradeId:
          type: integer
        journalEntryId:
          type: ["integer", "null"]
        asset:
          type: string
        entryDate:
          type: string
        entryPrice:
          type: number
        exitPrice:
          type: number
        direction:
          type: string
          enum: [long, short]
        marketContext:
          type: string
        newsEvents:
          type: array
          items:
            type: string
        sentimentConditions:
          type: string
        whaleActivity:
          type: string
        technicalIndicators:
          type: object
          properties:
            rsi:
              type: number
            macd:
              type: string
            trend:
              type: string
        userReasoning:
          type: string
        aiRecommendation:
          type: string
          enum: [buy, sell, wait, hold]
        actualOutcome:
          type: string
        lessonsLearned:
          type: string
        pnl:
          type: number
        pnlPercent:
          type: number
        behaviorFlags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ReplayRequest:
      type: object
      required: [tradeId]
      properties:
        tradeId:
          type: integer

    # ── Chat ─────────────────────────────────────────────────────
...